+++ app/src/main/java/com/socialite/solite_pos/data/source/local/room/ProductsDao.kt
@@ -15,8 +15,8 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface ProductsDao {
 
-    @Query("SELECT * FROM ${Product.DB_NAME} WHERE ${AppDatabase.UPLOAD} = 0")
-    suspend fun getNeedUploadProducts(): List<Product>
+    @Query("SELECT * FROM ${NewProduct.DB_NAME} WHERE ${AppDatabase.UPLOAD} = 0")
+    suspend fun getNeedUploadProducts(): List<NewProduct>
 
     @Query("SELECT * FROM ${NewProduct.DB_NAME} WHERE ${NewProduct.ID} = :idProduct LIMIT 1")
     suspend fun getProduct(idProduct: String): NewProduct
@@ -42,9 +42,6 @@ interface ProductsDao {
     @Query("SELECT * FROM ${NewProduct.DB_NAME}")
     fun getAllProductWithCategories(): Flow<List<ProductWithCategory>>
 
-    @Query("SELECT * FROM '${Product.DB_NAME}'")
-    suspend fun getProducts(): List<Product>
-
     @Query("UPDATE ${Product.DB_NAME} SET ${Product.STOCK} = ((SELECT ${Product.STOCK} FROM ${Product.DB_NAME} WHERE ${Product.ID} = :idProduct) + :amount) WHERE ${Product.ID} = :idProduct")
     fun increaseProductStock(idProduct: Long, amount: Int)
 
@@ -58,7 +55,7 @@ interface ProductsDao {
     suspend fun insertNewProduct(data: NewProduct)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProducts(list: List<Product>)
+    suspend fun insertProducts(list: List<NewProduct>)
 
     @Update
     suspend fun updateProduct(data: Product)
