+++ app/src/main/java/com/socialite/solite_pos/data/source/remote/response/entity/ProductResponse.kt
@@ -1,33 +1,25 @@
 package com.socialite.solite_pos.data.source.remote.response.entity
 
-import com.socialite.solite_pos.data.source.local.entity.room.master.Product
+import com.socialite.solite_pos.data.source.local.entity.room.new_master.Product
 
 data class ProductResponse(
-    val buyPrice: Int,
-    val category: Int,
+    val category: String,
     val desc: String,
     val id: String,
     val image: String,
     val isActive: Boolean,
-    val isMix: Boolean,
     val isUploaded: Boolean,
     val name: String,
-    val portion: Int,
-    val sellPrice: Int,
-    val stock: Int
+    val price: Int,
 ) {
     fun toEntity(): Product {
         return Product(
-            id = id.toLong(),
+            id = id,
             name = name,
-            category = category.toLong(),
+            category = category,
             image = image,
             desc = desc,
-            sellPrice = sellPrice.toLong(),
-            buyPrice = buyPrice.toLong(),
-            portion = portion,
-            stock = stock.toLong(),
-            isMix = isMix,
+            price = price.toLong(),
             isActive = isActive,
             isUploaded = isUploaded
         )
